Change key structure used in housekeeping state
HousekeepingManager keeps state that track of goroutines that are
running housekeeping against various repositories to prevent them
from concurrently doing so on a given repository. It's currently
using the absolute path of a repository to do so by calling `Path()`.
This is problematic as it leads to disk access, and this is done
before the manager sets up a transaction. Repositories can be uniquely
identified by `(storage, relative_path)`, and that doesn't entail
any disk access either. Let's use that as the key instead.