Extract transaction starting logic into a separate method
optimizeRepositoryWithTransaction is currently starting a
transaction always. We need to do this conditionally soon. RPCs
have transactions started already in the middleware where as
OptimizeRepository called outside of an RPC context needs to start
its own transaction. Extract the transaction starting logic into
a separate method so it is easier to change it.

Tests using a custom optimizeFunc were not running with transactions
before as the custom optimizeFunc also stubbed out the transaction
starting. Now they also run with transactions.