Start transaction for OptimizeRepository in middleware
OptimizeRepository has been special cased to start its own
transaction. The custom logic had issues earlier that led to
doing disk operations outside of a transaction. Start a transaction
for it in the middleware as we do for all other RPCs. This ensures
the RPC always operates in a transaction and doesn't accidentally
perform non-transactional disk operations.

To facilitate this, OptimizeRepository gains a configuration option
that has it use the existing transaction that is open rather than
starting its own one.