raft: Implement an abstraction to interact with Raft cluster
Although dragonboat provides an intensive set of interfaces to interact
with a Raft cluster, it's quite verbose to use those interfaces
directly. They might also become a foot gun if the caller is not mindful
of their implications. Thus, this commit implements an abstraction that
allows the caller to interact with the cluster safely and restrictively.

This abstraction uses generics to enforce request/response types. It
also standardizes the update result type of the state machine.