Let make notice-up-to-date use diff instead of cmp
make notice-up-to-date command is used to compare auto-generated NOTICE
file against the counterpart attached in the source code. It uses `cmp`
command for the comparison. As the file becomes enormous, that command
starts to report subtle changes. Those changes are not detectable by
`git`. I suppose they are platform dependent. This commit replaces that
command by `git-diff`. The new command aligns better with Gitaly and
helps us avoid frustrations when adding new dependency.