Replace CallLogManager with AccessLogManager
The `CallLogManager` function to the `LogManagerAccessor` interface
added in 2414ff76d (Add the LogManagerAccessor interface, 2024-06-07)
is somewhat confusing, and may cause a partition to open and close in
rapid succession.

Replace it with `AccessLogManager`, which returns a `LogManager` and a
close function that the caller must invoke when finished. This allows
the `LogConsumer` to keep the partition open for the full duration of
its operation.