Add a test for concurrent repacking and alternate linking
This commit adds a test case covering a scenario with concurrent
alternate unlinking and repacking operation. Repacking transactions
remove loose objects. Generally this is fine as new objects written
by transactions are always in packs. The only exception is alternate
unlinking operations which directly link the objects from the alternate
into the target repository. If some of these are loose objects, a
repacking operation removing loose objects may leave the repository in
a bad state.