Prevent loose object pruning with concurrent alternate unlinking
Applying a repacking operation prunes all loose objects on application.
If loose objects were concurrently introduced in the repository with
the repacking operation, this could lead to corruption if we prune a
loose object that is needed. Transactions in general only introduce
packs, not loose objects. The only exception to this currently is alternate
unlinking operations where the objects of the alternate are hard linked
into the member repository. This can technically still introduce loose
objects into the repository and trigger this problem as the pools could
still have loose objects in them until the first repack.

Check if the repository was unlinked from an alternate concurrently and
abort the repacking if so.