storagemgr: Convert `InitialValues` to use `git.Reference`
The `InitialValues` in `RecordInitialReferenceValues` uses
`git.ObjectID` to denote the initial value for a reference. This
currently works since we only deal with regular references. We want to
move to support symbolic references too, since this will be used in the
context of reftables, where updating the default branch ('HEAD'
reference) is treated similar to a regular ref update.