raft: Fanout operations in Raft tests
Raft tests need to assert the states in each node. Recently, all of
those operations have been done serially. It slows down the test, as
well as decreases the chance of detecting racing bugs in the
distributed system. This commit adds some helper functions to fan-out
operations in the tests. As a result, it speeds up the Raft test suite.