Export ReadAlternatesFile
We're soon going to move the snapshotting logic into a separate
package. As it also depends on readAlternatesFile, export the function
so we can reuse it from the snapshotting package.

The exported function is moved to a new package due to cyclic imports.
'git' package import 'storage' so we can't place it there. The function
performs validation on the alternates that doesn't really fit the Git
package.