Convert Snapshot into an interface
We're soon going to add a second implementation of a snapshot that
allows for sharing a snapshot between multiple transactions. In
preparation for doing so, make the current snapshot type private and
export a public interface both of the types implement.

snapshot.FileSystem also stutters less as a name than snapshot.Snapshot.