featureflag: Make SetAttrTreeConfig flip at random
The `SetAttrTreeConfig` flag is set to always be 'false'. This is
ineffective as this feature captures a breaking change in Git. So let's
randomly enable it on a global level. Also fix a test which was
specifically flipping the value of the flag to no longer do that, since
we do it on a global level.