diff: Modify `TestDiffBlobs` to acknowledge `SetAttrTreeConfig`
The `TestDiffBlobs` tests assume that the 'attr.tree' config is set to
'HEAD'. While this is true in Git versions < 2.46.0, this patch was
reverted in master and will land in 2.46.0.

This means it will no longer be default. We have thus created the
`SetAttrTreeConfig` to capture this behavior. But this flag value is not
used in `TestDiffBlobs`, which sets some attributes. Modify the test to
acknowledge this flag and change the expected response correspondingly.