localrepo: Modify `TestRepo_ObjectHash` to acknowledge `SetAttrTreeConfig`
The test in `TestRepo_ObjectHash` checks how many times the 'git
rev-parse --show-object-format' is called to get the hash of the object,
since its cached after the first call.

When the `SetAttrTreeConfig` flag is set however, we also make a call
internally to check and set an empty tree ID for `SetAttrTreeConfig`.
Let's modify the test to acknowledge this.