Decouple global metrics from transaction and transaction manager
There's currently a metrics struct that contains housekeeping metrics
passed down to transcation manager and transaction. These metrics are
not scoped by the storage and are measured everywhere as is.

Introduce specific metric types for transaction manager and transaction.
This allows us to scope the metrics by the type better. For example, the
metrics passed into transaction manager could already be scoped by the
correct storage to ensure we don't have to drill down that information
into the transaction manager.