raft: Implement cluster bootstrapping
This commit implements Raft cluster bootstrapping to Gitaly. This
process boils down to 4 steps:

0. Initializing the node-level Raft object for each storage. It
initializes underlying engines, networking Raft servers, log
databases, etc.
1. Joining the metadata Raft group. All initial nodes are blocked
until the quorum is reached. A node member takes the leader role
of the metadata Raft group.
2. Bootstrapping the Raft cluster if configured to do so. The leader
of the metadata Raft group initializes cluster information and
replicates it to other nodes. If a node joins the metadata group late,
it gets the replicated info from the cluster.
3. Each node registers the node's storage with the metadata Raft group
in a random fashion. The metadata Raft group allocates a new storage
ID for each of them. They persist in their IDs. This type of ID is
used for future interaction with the cluster. A node enters "ready"
state afterward.

Gitaly bootstraps the cluster whenever it receives
`--init-raft-cluster` flag in `gitaly serve` command. If a node
restarts, it re-used the persisted information to connect to the
cluster. The cluster also performs some validation to ensure a node
joins the correct cluster and the storage is globally unique.