Fix flaky *ReceivePack tests
*ReceivePack_success tests are flaky at the moment because the
MockManager used in the tests is not fully consuming the stdin
stream before returning. This leads to the HookService RPCs returning
before `gitaly-hooks` has finished sending the stdin causing it to
fail with `error when receiving data for pre-receive hook: stdin send
error: EOF`. Fix the issue by consuming the stdin stream fully.