backoff: Add mutex to protect random source
From the doc (https://pkg.go.dev/math/rand#NewSource), the source of
rand package is not thread-safe. It's recommended to create
per-goroutine sources. That approach is super inconvenient. This commit
adds a mutex to protect rand source. This addition does not yield any
performance implications. Typically, the users of `backoff` package put
the goroutine to sleep for a certain amount of time before retrying.
Thus, mutex overhead is multiple orders of magnitude smaller than the
deliberate sleep.