raft: Upgrade cockroachdb/pebble package
The recent version does not work with 32-bit platforms. This commit
bumps that package version to higher version. The bug was fixed in:
https://github.com/cockroachdb/pebble/issues/2853.

As the latest `pebble` version is not compatible with `dragonboat`.
There are some minor changes in `pebble`. So, we have no choice but to
fork `dragonboat` to upgrade the dependency:
https://gitlab.com/gl-gitaly/dragonboat/-/commit/725bd577d6ec83a7619edbf889af6e8195644ce9

We'll upstream this dependency bump later.