Extend UserMergeBranch with squash option
Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/355613

**Problem**

We want to support a new "Squash and merge" merge strategy. It's already
possible to squash commits during the merge, but it will lead to
creation of the merge commit.

"Squash and merge" strategy is going to squash all merge request commits
and merge only the resulted squash commit (without a merge commit).

**Solution**

Gitaly `merge` operation already supports a `squash` option, but it's
not exposed via `UserMergeBranch` RPC. This commit extends
`UserMergeBranch` to accept a `squash` option in
`UserMergeBranchRequest` payload.

Changelog: added