diff: Add patch limits to `DiffBlobs` RPC
The `DiffBlobs` RPC generates batches of diffs from pairs of blob
revisions. Since the client must provide the blob revision pairs, it
controls how many diffs are too be generated in the batch. If the size
of the blobs being diffed is very large, the diff also be very large and
not worth transmitting back to the client.

Add `patch_bytes_limit` to the `DiffBlobs` RPC request to provide
clients with a way to specify the maximum size of a single patch that
should be returned. Patches exceeding this limit do not set any patch
data and set `over_patch_bytes_limit` in the response.