Disable fsyncing by Git with transactions
When transactions are enabled, the Git commands execute against
the transaction's snapshot. These writes don't need to be persisted
in the snapshot as the TransactionManager will write them into the
write-ahead log and fsync the WAL entry. It is thus wasteful for
Git to fsync its writes that are ultimately discarded anyway.

Disable fsyncing by Git and rely on the TransactionManager to fsync
its writes.