Enforce directory permissions in storage
TransactionManager is currently using the permissions set on the
directory by the transaction. This in general is okay but can lead
to various permissions in the storage. It's also possible for a
transaction to create a directory hierarchy and mark it read-only
before comitting it. This would lead to log application failures
as the child entries in the directory can't be created as we'd log
the directory also with read-only permissions.

Storage permissions shouldn't really be managed by anything else
except for the storage layer. This commit starts ignoring the
permissions on the directory and stores all directories in the
storage with +rwx permissions for the owning user only.