Enforce file permissions in storage
We're currently storing files with whatever permissions the transaction
set for them. This can lead to storing files with bad permissions that
don't give Gitaly access to read them. More generally, the storage
layer should be the only component concerned with file permissions.
This commit standardizes all files to be stored as either readable by
the owner, or readable and executable by the owner. This ensures only
Gitaly can read the files, and that all files are stored with consistent
permissions.