Ensure page token is for the same tree entry
Contributes to https://gitlab.com/gitlab-org/gitaly/-/issues/4556

The page token returned does not include the tree oid and can
page across two different trees leading to incorrect results.
In order to resolve this we can include the tree oid in the page token and
use it to ensure the page token is always used to access a given tree only.

Changelog: changed