Remove GroupPrivateDir permission
GroupPrivateDir permission is only used when creating the parent
directories of a lock file when creating a repository. In practice
this means the '@hashed/xx/yy' hierarchy. There doesn't seem to be
a real reason why the directory hierarchy should be readable for
anyone except Gitaly's user. Harmonize this to use the PrivateDir
permission.