Remove PublicDir permission
Outside of tests, PublicDirs permission is used in two locations:

1. Quarantine directories are marked PublicDir.
2. `custom_hooks/` directories are marked public dir.

Quarantines are internal runtime state of Gitaly and don't need to
be readable by others. Custom hooks directories in the repositories
also shouldn't need to be accessible by anything but Gitaly.

The remaining use of PublicDir permission is in tests.

Replace all of the PublicDir usage with PrivateDir.