Fix flaky test TestGetSpawnToken_CommandStats
TestGetSpawnToken_CommandStats is flaky as it has a ten second
timeout on acquiring a spawn token. In slower execution, for
example in the CI, this timeout can be reached and the test flake.

There's no real reason to set a separate timeout here. We can rely
on the general testing timeout after which the testing run ends.
Remove the timeout. Additionally, replace the sleep with a call
to Gosched to achieve the result we're after without sleeping.